requirements.txt
GitPython==3.1.29
python_sonarqube_api==1.3.1